requirements.txt
pytest
pytest_bdd
selenium
jsonpickle
pymysql
mysql-connector-python-rf
pony
pytest-allure-adaptor